requirements.txt
# 網頁爬蟲相關
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Playwright 支援（保留 Selenium 作為備選）
selenium==4.15.2

# 資料處理
pandas==2.1.3
numpy==1.25.2

# 配置管理
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# 日誌和監控
loguru==0.7.2

# 測試
pytest==7.4.3
pytest-asyncio==0.21.1

# 開發工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1